+++ lib/convert.d.ts
@@ -1,1 +1,1 @@
-export { e as ConvertOptions, f as ConvertRequest, g as ConvertResult, R as ResolvedUnit, h as UnitConversionError, k as convert, l as convertDetailed, m as convertRaw, p as parseRequest, r as resolveUnit, q as roundSignificant } from './convert-D6txqQbW.js';
+export { e as ConvertOptions, f as ConvertRequest, g as ConvertResult, R as ResolvedUnit, h as UnitConversionError, k as convert, l as convertDetailed, m as convertRaw, p as parseRequest, r as resolveUnit, q as roundSignificant } from './convert-DVefQ0C-.js';